+++ penman/lexer.py
@@ -22,14 +22,12 @@ PATTERNS = {
     'STRING':     r'"[^"\\]*(?:\\.[^"\\]*)*"',
     # ROLE cannot be made up of COLON + SYMBOL because it then becomes
     # difficult to detect anonymous roles: (a : b) vs (a :b c)
-    'ROLE':       r':[^\s()\/:~,]*',
-    'SYMBOL':     r'[^\s()\/:~,]+',
+    'ROLE':       r':[^\s()\/:~]*',
+    'SYMBOL':     r'[^\s()\/:~]+',
     'ALIGNMENT':  r'~(?:[a-zA-Z]\.?)?\d+(?:,\d+)*',
     'LPAREN':     r'\(',
     'RPAREN':     r'\)',
     'SLASH':      r'\/',  # concept (node label) role
-    'COMMA':      r',',   # used in triple conjunctions
-    'CARET':      r'\^',  # used in triple conjunctions
     'UNEXPECTED': r'[^\s]'
 }
 
@@ -50,7 +48,7 @@ PENMAN_RE = _compile('COMMENT',
                      'UNEXPECTED')
 TRIPLE_RE = _compile('COMMENT',
                      'STRING',
-                     'LPAREN', 'RPAREN', 'COMMA', 'CARET',
+                     'LPAREN', 'RPAREN',
                      'SYMBOL',
                      'UNEXPECTED')